test(cuda.core): share PDL overlap protocol via helper runner
Extract kernels and the same-stream / graph-capture overlap check into
run_pdl_overlap_check so launcher and GraphBuilder tests stay in sync.
