+++ cuda_core/tests/graph/test_graph_builder.py
@@ -7,15 +7,15 @@ import gc
 import time
 import weakref
 
-import helpers
 import numpy as np
 import pytest
 from conftest import skipif_need_cuda_headers
 from cuda_python_test_helpers.marks import requires_module
 from helpers.graph_kernels import compile_common_kernels, compile_conditional_kernels
 from helpers.misc import try_create_condition
+from helpers.pdl_kernels import run_pdl_overlap_check
 
-from cuda.core import Device, LaunchConfig, LegacyPinnedMemoryResource, Program, ProgramOptions, launch
+from cuda.core import Device, LaunchConfig, LegacyPinnedMemoryResource, launch
 from cuda.core.graph import GraphBuilder, GraphDefinition
 from cuda.core.graph._graph_builder import (
     _capture_callback_with_tail_failure_for_testing,
@@ -783,85 +783,4 @@ def test_pdl_primary_secondary_overlap_graph_capture(init_cuda):
     but launches are captured into a CUDA graph (see CUDA Programming Guide,
     Programmatic Dependent Launch). Overlap is opportunistic → miss is xfail.
     """
-    dev = Device()
-    if dev.compute_capability < (9, 0):
-        pytest.skip("Programmatic Dependent Launch requires compute capability >= 9.0")
-    stream = dev.create_stream(options={"nonblocking": True})
-
-    # clock64 budgets are in GPU cycles; keep the post-trigger window long enough
-    # for the secondary to boot, but short enough for a unit test.
-    code = r"""
-    #include <cuda_device_runtime_api.h>
-
-    extern "C" __global__ void primary_kernel(int* secondary_started, int* overlapped) {
-        cudaTriggerProgrammaticLaunchCompletion();
-
-        const long long deadline = clock64() + 100000000LL;  // ~50ms @ ~2GHz
-        if (threadIdx.x == 0 && blockIdx.x == 0) {
-            while (clock64() < deadline) {
-                if (atomicAdd(secondary_started, 0) != 0) {
-                    atomicExch(overlapped, 1);
-                    return;
-                }
-                __nanosleep(1000);
-            }
-        }
-    }
-
-    extern "C" __global__ void secondary_kernel(int* secondary_started) {
-        if (threadIdx.x == 0 && blockIdx.x == 0) {
-            atomicExch(secondary_started, 1);
-        }
-    }
-    """
-
-    arch = "".join(f"{i}" for i in dev.compute_capability)
-    pro_opts = ProgramOptions(std="c++17", arch=f"sm_{arch}", include_path=helpers.CUDA_INCLUDE_PATH)
-    prog = Program(code, code_type="c++", options=pro_opts)
-    mod = prog.compile("cubin")
-    primary = mod.get_kernel("primary_kernel")
-    secondary = mod.get_kernel("secondary_kernel")
-
-    mr = LegacyPinnedMemoryResource()
-    secondary_started = np.from_dlpack(mr.allocate(4)).view(np.int32)
-    overlapped = np.from_dlpack(mr.allocate(4)).view(np.int32)
-
-    primary_cfg = LaunchConfig(grid=1, block=1)
-    secondary_cfg = LaunchConfig(grid=1, block=1, programmatic_stream_serialization=True)
-    secondary_serial_cfg = LaunchConfig(grid=1, block=1)
-
-    def _run(secondary_launch_cfg: LaunchConfig) -> int:
-        secondary_started[0] = 0
-        overlapped[0] = 0
-        gb = stream.create_graph_builder().begin_building()
-        launch(gb, primary_cfg, primary, secondary_started.ctypes.data, overlapped.ctypes.data)
-        launch(gb, secondary_launch_cfg, secondary, secondary_started.ctypes.data)
-        graph = gb.end_building().complete()
-        try:
-            graph.launch(stream)
-            stream.sync()
-        finally:
-            graph.close()
-            gb.close()
-        return int(overlapped[0])
-
-    # Without the PDL attribute, same-stream kernels stay serialized even via graph.
-    assert _run(secondary_serial_cfg) == 0, "Expected no overlap when programmatic_stream_serialization is False"
-
-    # PDL overlap is opportunistic; retry a few times on a quiet GPU.
-    saw_overlap = False
-    for _ in range(5):
-        if _run(secondary_cfg) == 1:
-            saw_overlap = True
-            break
-
-    if not saw_overlap:
-        pytest.xfail(
-            "PDL (Programmatic Dependent Launch) graph-capture overlap was not observed. "
-            "If this keeps xfailing in CI, manually re-check on a quiet Hopper+ GPU."
-        )
-
-    print(
-        f"PDL graph-capture overlap verified on {dev.name} compute capability {dev.compute_capability}",
-        flush=True,
-    )
+    run_pdl_overlap_check(Device(), via_graph=True)
